Django/Video_Project/requirements.txt
Django==2.2.3
django-redis==4.10.0
django-redis-cache==2.0.0
mysqlclient==1.4.2.post1
Pillow==6.1.0
PyMySQL==0.9.3
pytz==2019.1
redis==3.3.5
sqlparse==0.3.0
